requirements.txt
PyYAML==6.0.1
rich==13.7.0
openai==1.3.5
python-dotenv==1.0.0
gitpython==3.1.40